requirements.txt
Django==2.1.3
Pillow==5.4.1
pytz==2018.9
